requirements/requirements-pipeline.txt
azureml-core==1.28.0
azureml-pipeline==1.28.0
azureml-sdk==1.28.0
azure-ml-component==0.1.0.36279725
cryptography==3.3.2
azure-cli-core==2.22.1
ruamel.yaml~=0.16
gitpython~=3.1
hydra-core==1.0.3 

# Telemetry logging
opencensus-ext-azure==1.0.7
